refactor(dao): lift internal/dao/interfaces to pkg/dao (#177)
