+++ internal/dao/memory/event_dao_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/i2-open/i2goSignals/pkg/dao/ids"
-	"github.com/i2-open/i2goSignals/internal/dao/interfaces"
+	interfaces "github.com/i2-open/i2goSignals/pkg/dao"
 	"github.com/i2-open/i2goSignals/pkg/goSet"
 	"github.com/i2-open/i2goSignals/pkg/ssfModels"
 )
