+++ internal/dao/memory/stream_dao.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/i2-open/i2goSignals/internal/dao/interfaces"
+	interfaces "github.com/i2-open/i2goSignals/pkg/dao"
 	"github.com/i2-open/i2goSignals/pkg/ssfModels"
 )
