+++ internal/dao/memory/subject_filter_dao_test.go
@@ -6,7 +6,7 @@ import (
     "testing"
     "time"
 
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     model "github.com/i2-open/i2goSignals/pkg/ssfModels"
 )
