+++ internal/dao/memory/token_dao.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/i2-open/i2goSignals/internal/dao/interfaces"
+	interfaces "github.com/i2-open/i2goSignals/pkg/dao"
 	"github.com/i2-open/i2goSignals/pkg/ssfModels"
 )
