+++ internal/dao/mongo/key_dao.go
@@ -6,7 +6,7 @@ import (
     "slices"
 
     "github.com/i2-open/i2goSignals/pkg/dao/ids"
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     "github.com/i2-open/i2goSignals/pkg/logger"
     "go.mongodb.org/mongo-driver/v2/bson"
     "go.mongodb.org/mongo-driver/v2/mongo"
