+++ internal/dao/mongo/server_dao.go
@@ -4,7 +4,7 @@ import (
     "context"
     "errors"
 
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     "github.com/i2-open/i2goSignals/pkg/logger"
     "github.com/i2-open/i2goSignals/pkg/ssfModels"
     "go.mongodb.org/mongo-driver/v2/bson"
