+++ internal/dao/mongo/subject_filter_dao_test.go
@@ -8,7 +8,7 @@ import (
     "testing"
     "time"
 
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     model "github.com/i2-open/i2goSignals/pkg/ssfModels"
     "github.com/stretchr/testify/suite"
     "go.mongodb.org/mongo-driver/v2/bson"
