+++ internal/providers/dbProviders/mongo_provider/test/rebind_test.go
@@ -8,7 +8,7 @@ import (
     "time"
 
     "github.com/i2-open/i2goSignals/pkg/dao/ids"
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     "github.com/i2-open/i2goSignals/internal/providers/dbProviders/mongo_provider"
     "github.com/stretchr/testify/assert"
     "github.com/stretchr/testify/suite"
