+++ internal/server/test/register_provenance_test.go
@@ -7,7 +7,7 @@ import (
     "testing"
     "time"
 
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     "github.com/i2-open/i2goSignals/internal/dao/memory"
     "github.com/i2-open/i2goSignals/internal/services"
     "github.com/i2-open/i2goSignals/pkg/authSupport"
