+++ internal/services/stream_service_subject_filter_test.go
@@ -8,7 +8,7 @@ import (
     "strings"
     "testing"
 
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     "github.com/i2-open/i2goSignals/internal/dao/memory"
     "github.com/i2-open/i2goSignals/pkg/ssfModels"
     "github.com/stretchr/testify/assert"
