+++ internal/services/subject_filter_service.go
@@ -5,7 +5,7 @@ import (
     "errors"
     "time"
 
-    "github.com/i2-open/i2goSignals/internal/dao/interfaces"
+    interfaces "github.com/i2-open/i2goSignals/pkg/dao"
     "github.com/i2-open/i2goSignals/pkg/goSet"
     model "github.com/i2-open/i2goSignals/pkg/ssfModels"
     "github.com/i2-open/i2goSignals/pkg/subjectid"
