+++ pkg/dao/dao_interfaces.go
@@ -1,4 +1,4 @@
-package interfaces
+package dao
 
 import (
     "context"
